Here's what we have so far, but I've realised it's wrong.
Sending Play and Place at the same time, trying to do this everywhere
and handling it only with the expectation that we'll have both at any
time, this is causing more trouble than it's worth, and it's not a
very Backboney way to do things. From here on out we need to break
the Backbone code into two, handle a Play and a Place one at a time.